feature WeeklyMenu endpoint implemented, testet with rest assured and integration db test all test green

+++ src/main/java/app/dtos/menu/WeeklyMenuOverviewDTO.java
@@ -10,7 +10,7 @@ public record WeeklyMenuOverviewDTO(
     int weekNumber,
     int year,
     MenuStatus menuStatus,
-    int slotCount,
+    Long slotCount,
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
     LocalDateTime publishedAt
 ) {}
